backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
click==8.2.0
fastapi==0.115.12
greenlet==3.2.2
gunicorn==23.0.0
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pydantic==2.11.4
pydantic_core==2.33.2
pyhumps==3.8.0
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
